HBP-90. Move methods to gather_references from ABCAdapter to h5.
