+++ framework_tvb/tvb/core/neocom/h5.py
@@ -34,6 +34,7 @@ import uuid
 
 from tvb.basic.neotraits.api import HasTraits
 from tvb.core.entities.generic_attributes import GenericAttributes
+from tvb.core.entities.load import load_entity_by_gid
 from tvb.core.entities.model.model_datatype import DataType
 from tvb.core.neocom._h5loader import Loader, DirLoader, TVBLoader
 from tvb.core.neocom._registry import Registry
@@ -256,3 +257,44 @@ def load_view_model_from_file(filepath):
                 loaded_ref = load_view_model(gid, base_dir)
             setattr(view_model, trait_attr.field_name, loaded_ref)
     return view_model
+
+
+def gather_all_references_by_index(h5_file, ref_files):
+    refs = h5_file.gather_references()
+    for _, gid in refs:
+        if not gid:
+            continue
+        index = load_entity_by_gid(gid.hex)
+        h5_file = h5_file_for_index(index)
+        ref_files.append(h5_file.path)
+        gather_all_references_by_index(h5_file, ref_files)
+
+
+def gather_all_references_of_view_model(gid, base_dir, ref_files):
+    vm_path = determine_filepath(gid, base_dir)
+    ref_files.append(vm_path)
+    view_model_class = H5File.determine_type(vm_path)
+    view_model = view_model_class()
+
+    with ViewModelH5(vm_path, view_model) as vm_h5:
+        references = vm_h5.gather_references()
+        uuids = vm_h5.gather_references_by_uuid()
+
+        for _, gid in references:
+            if not gid:
+                continue
+            if isinstance(gid, (list, tuple)):
+                for list_gid in gid:
+                    gather_all_references_of_view_model(list_gid, base_dir, ref_files)
+            else:
+                gather_all_references_of_view_model(gid, base_dir, ref_files)
+
+        uuid_files = []
+        for _, gid in uuids:
+            if not gid:
+                continue
+            index = load_entity_by_gid(gid.hex)
+            h5_file = h5_file_for_index(index)
+            uuid_files.append(h5_file.path)
+            gather_all_references_by_index(h5_file, uuid_files)
+        ref_files.extend(uuid_files)
